Created a separate API for time tracking
